Feature 20240808 rule jobhandler

See merge request pd_str/marketing!1011

+++ marketing-cs/src/main/java/com/br/marketing/mapper/CustomerInfoPushMainMapperBase.java
@@ -2,11 +2,12 @@ package com.br.marketing.mapper;
 
 import com.br.marketing.entity.CustomerInfoPushMain;
 import com.br.marketing.entity.CustomerInfoPushMainExample;
-import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
+import java.util.List;
+
 public interface CustomerInfoPushMainMapperBase {
-    long countByExample(CustomerInfoPushMainExample example);
+    int countByExample(CustomerInfoPushMainExample example);
 
     int deleteByExample(CustomerInfoPushMainExample example);
